+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2021-03-09
+
 ### Added
 - portal-316: Added `unioncloud:members:status` and `unioncloud:usermemberships:status` commands to monitor the progress of the caching.
 
@@ -18,5 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/bristol-su/unioncloud-portal/compare/v1.0.0...HEAD 
+[Unreleased]: https://github.com/bristol-su/unioncloud-portal/compare/v1.0.1...HEAD 
+[1.0.1]: https://github.com/bristol-su/unioncloud-portal/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/bristol-su/unioncloud-portal/releases/tag/v1.0.0